Backend/requirements.txt
python-telegram-bot[job-queue]>=22.6
openai>=1.30.1
groq>=0.9.0
apscheduler==3.10.4
dateparser>=1.2.0
python-dotenv>=1.0.1
aiosqlite>=0.20.0
httpx>=0.27.0
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
pydantic>=2.0.0
sqlite-vec>=0.1.6
pinecone>=5.0.0
psutil>=5.9.0

# Dev / тесты
pytest>=8.0.0
pytest-asyncio>=1.0.0
httpx>=0.27.0
